Extract to two separate functions for iso8601 and timestamps
